Refactor NeoFaker.Boolean
